Expose internal utils
Rename rmColSpan to removeColSpan
