+++ src/util.js
@@ -3,6 +3,8 @@
 import {PluginKey} from "prosemirror-state"
 
 import {TableMap} from "./tablemap"
+import {tableNodeTypes} from "./schema";
+import {CellSelection} from "./cellselection";
 
 export const key = new PluginKey("selectingCells")
 
@@ -80,7 +82,7 @@ export function setAttr(attrs, name, value) {
   return result
 }
 
-export function rmColSpan(attrs, pos, n=1) {
+export function removeColSpan(attrs, pos, n=1) {
   let result = setAttr(attrs, "colspan", attrs.colspan - n)
   if (result.colwidth) {
     result.colwidth = result.colwidth.slice()
@@ -97,4 +99,30 @@ export function addColSpan(attrs, pos, n=1) {
     for (let i = 0; i < n; i++) result.colwidth.splice(pos, 0, 0)
   }
   return result
+}
+
+export function columnIsHeader(map, table, col) {
+  let headerCell = tableNodeTypes(table.type.schema).header_cell
+  for (let row = 0; row < map.height; row++)
+    if (table.nodeAt(map.map[col + row * map.width]).type != headerCell)
+      return false
+  return true
+}
+
+
+// Helper to get the selected rectangle in a table, if any. Adds table
+// map, table node, and table start offset to the object for
+// convenience.
+export function selectedRect(state) {
+  let sel = state.selection, $pos = selectionCell(state)
+  let table = $pos.node(-1), tableStart = $pos.start(-1), map = TableMap.get(table)
+  let rect
+  if (sel instanceof CellSelection)
+    rect = map.rectBetween(sel.$anchorCell.pos - tableStart, sel.$headCell.pos - tableStart)
+  else
+    rect = map.findCell($pos.pos - tableStart)
+  rect.tableStart = tableStart
+  rect.map = map
+  rect.table = table
+  return rect
 }
